third_party/WebKit/LayoutTests/editing/execCommand/selection-after-insert-list.html

<script src="../../resources/dump-as-markup.js"></script>
<div id="test" contenteditable>
<ul>
    <li id="first">ABC</li>
    <li id="second">DEF</li>
</ul>
</div>
<script>
var selection = window.getSelection();
var range = document.createRange();
range.setStart(document.getElementById('first').firstChild, 2);
range.setEnd(document.getElementById('second').firstChild, 3);
selection.addRange(range);
document.execCommand('InsertOrderedList', false, null);
Markup.description("This tests verifies that the selection is properly reset after executing the command.");
Markup.dump('test');
</script>
