third_party/WebKit/LayoutTests/editing/execCommand/selection-after-insert-list-expected.txt

This tests verifies that the selection is properly reset after executing the command.
| <ol>
|   <li>
|     "AB<#selection-anchor>C"
|   <li>
|     id="second"
|     "DEF<#selection-focus>"
| "
"
